Prefer JAVA_HOME over java from PATH if set
In case the environment variable JAVA_HOME is set and is considered
somewhat sane (it contains bin/java) prefer this java binary over the
one we might find in the PATH.

For this an additional variable JAVA is introduced, which contains the
java binary to call. It will contain the absolute path to java from
JAVA_HOME or just 'java' in case JAVA_HOME is not set; which gives the
same behaviour as before.