Run custom scripts (#1094)
Use --browsertime.script myScript.js to add your script. You can run multiple script by passing the parameter multiple times.

The metrics will automatically turn up on the summary page, detailed page, the summary page of the page and the run page. They will also be sent to Graphite.

We lack support for passing a dir with multiple files (as we have in Browsertime) but I think it's ok to wait with that! #1063