Store analysis per group (#1725)
* Avoid name collisions for stored analysis files.

Summary messages can be different based on group (most often per domain). Include group in file name, if present to avoid overwriting data in an existing file.

* Remove redundant summary messages.

The ”largestassets.summary” message with group:’total’ was sent with different structure for message data from what was sent earlier in this function.

* Additional validation of summary and pageSummary.

Check that url and group are specified correctly for all messages.

* Use ES6 template strings instead of util.format.

* Warn about duplicated summary messages.

This will be an error later, but need to update the assets plugin first.

* Only put summary data on index and detailed page.

Summaries with group=’total’ is data for the entire test (i.e. all url and domains). Previously the last group message (which might be just a subset of all data) would be used.

* No duplicate largestassets.summary messages.

Change so there’s only one largestassets.summary message per group, with data for all content types. Then enforce this for all summary messages.