Switch from hooks to messages
This is the first step to get rid off hooks:
* Introduce two new messages: setup and final
* Logs a warning messages for all plugins that uses old hooks

The setup message will make it possible for plugins to pass
configuration to each other (send pugs to the HTML plugin,
send javascript (that collects metrics) to Browsertime and can
hopefully remove the post/preOpen hooks. The final message
replaces the postClose and final hook (or do we need more messages?).

Keep the hoosk deperecated through 6 and remove them in 7.0.