Tag CPU data and Lighthouse score in InfluxDB (#2225)
* Add specific tags for CPU data in InfluxDB

Pre this version you could get the CPU data but it was kind of hard
to understand the tags. With this fix we add a cpu tag that either is
category or event. That way it's easier to pickup (and understand)
that data.

* Add audit tag for Lighthouse tests sent to InfluxDB

Before the fix we just sent "score" but no tag telling which score.
We now add a tag named audit that tells which audit that is used
for the score.

See #2222 and #2224