fix(plugins/slack): correctly send messages on html.finished message
Do not rely on `resultBaseUrl` to block messages to be sent to Slack. Instead, check if any
of the supported hosting providers are correctly configured. If that is the case, then
messages are sent _only after the upload is complete_. For all other cases, send the message
anyway

test(plugins/slack): add basic test scenarios

- Check for missing options when calling `slack.open()`
- Ensure that methods supposed to return lists always return them
- Add some sanity checks to the summary, so that missing values would not crash