Fix HTML report when using "browsertime.chrome.visualMetricsUsingTrace"
The "browsertime.chrome.visualMetricsUsingTrace" option is a great
way of speeding up test runs, but it causes some problems with the HTML
report. For example, the "pages.pug" template won't show table headers
for the visual metrics gathered, yet they're present in the table rows,
causing a mismatch between the header and rows.

This fixes the affected templates to recognize when
"browsertime.chrome.visualMetricsUsingTrace" is enabled.