Fix browserScript structure.
Oh, we had a really strange behaviour where we added all the Browsertime data
inside the browserScript element (the data we collect with JS) and that
made it so we pushed the HAR file inside that element. This fixes that and make
sure we just cherry pick the exact data we pass on.